@@ -1,18 +1,23 @@
 {
   "name": "@angular/service-worker",
-  "version": "5.1.0-beta.0-21bfaf226",
+  "version": "7.0.0-beta.4-a2418a9037",
   "description": "Angular - service worker tooling!",
   "main": "./bundles/service-worker.umd.js",
-  "module": "./esm5/service-worker.js",
-  "es2015": "./esm15/service-worker.js",
+  "module": "./fesm5/service-worker.js",
+  "es2015": "./fesm2015/service-worker.js",
+  "esm5": "./esm5/service-worker.js",
+  "esm2015": "./esm2015/service-worker.js",
+  "fesm5": "./fesm5/service-worker.js",
+  "fesm2015": "./fesm2015/service-worker.js",
   "typings": "./service-worker.d.ts",
   "author": "angular",
   "license": "MIT",
   "dependencies": {
-    "tslib": "^1.7.1"
+    "tslib": "^1.9.0"
   },
   "peerDependencies": {
-    "@angular/core": "5.1.0-beta.0-21bfaf226"
+    "@angular/core": "7.0.0-beta.4-a2418a9037",
+    "@angular/common": "7.0.0-beta.4-a2418a9037"
   },
   "repository": {
     "type": "git",
@@ -20,5 +25,9 @@
   },
   "bin": {
     "ngsw-config": "./ngsw-config.js"
-  }
+  },
+  "ng-update": {
+    "packageGroup": ["@angular/compiler", "@angular/core", "@angular/common", "@angular/animations", "@angular/platform-browser", "@angular/platform-browser-dynamic", "@angular/forms", "@angular/http", "@angular/platform-server", "@angular/platform-webworker", "@angular/platform-webworker-dynamic", "@angular/upgrade", "@angular/router", "@angular/compiler-cli", "@angular/language-service", "@angular/benchpress", "@angular/service-worker", "@angular/elements"]
+  },
+  "sideEffects": false
 }
