@@ -1,3 +1,10 @@
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
 import { Filesystem } from './filesystem';
 import { Config } from './in';
 /**
@@ -10,6 +17,10 @@ export declare class Generator {
     private baseHref;
     constructor(fs: Filesystem, baseHref: string);
     process(config: Config): Promise<Object>;
-    private processAssetGroups(config, hashTable);
-    private processDataGroups(config);
+    private processAssetGroups;
+    private processDataGroups;
 }
+export declare function processNavigationUrls(baseHref: string, urls?: string[]): {
+    positive: boolean;
+    regex: string;
+}[];
