@@ -1,1 +1,8 @@
-export declare function globToRegex(glob: string): string;
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+export declare function globToRegex(glob: string, literalQuestionMark?: boolean): string;
