@@ -1,11 +1,13 @@
 /**
- * @license Angular v5.1.0-beta.0-21bfaf226
- * (c) 2010-2017 Google, Inc. https://angular.io/
+ * @license Angular v7.0.0-beta.4-a2418a9037
+ * (c) 2010-2018 Google, Inc. https://angular.io/
  * License: MIT
  */
+import { __awaiter } from 'tslib';
+
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes} checked by tsc
+ * @suppress {checkTypes,extraRequire,uselessCode} checked by tsc
  */
 /**
  * @license
@@ -14,25 +16,31 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
+/** @type {?} */
 const PARSE_TO_PAIRS = /([0-9]+[^0-9]+)/g;
+/** @type {?} */
 const PAIR_SPLIT = /^([0-9]+)([dhmsu]+)$/;
 /**
  * @param {?} duration
  * @return {?}
  */
 function parseDurationToMs(duration) {
-    const /** @type {?} */ matches = [];
-    let /** @type {?} */ array;
+    /** @type {?} */
+    const matches = [];
+    /** @type {?} */
+    let array;
     while ((array = PARSE_TO_PAIRS.exec(duration)) !== null) {
         matches.push(array[0]);
     }
     return matches
         .map(match => {
-        const /** @type {?} */ res = PAIR_SPLIT.exec(match);
+        /** @type {?} */
+        const res = PAIR_SPLIT.exec(match);
         if (res === null) {
             throw new Error(`Not a valid duration: ${match}`);
         }
-        let /** @type {?} */ factor = 0;
+        /** @type {?} */
+        let factor = 0;
         switch (res[2]) {
             case 'd':
                 factor = 86400000;
@@ -59,7 +67,7 @@ function parseDurationToMs(duration) {
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes} checked by tsc
+ * @suppress {checkTypes,extraRequire,uselessCode} checked by tsc
  */
 /**
  * @license
@@ -68,23 +76,43 @@ function parseDurationToMs(duration) {
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-const WILD_SINGLE = '[^\\/]+';
+/** @type {?} */
+const QUESTION_MARK = '[^/]';
+/** @type {?} */
+const WILD_SINGLE = '[^/]*';
+/** @type {?} */
 const WILD_OPEN = '(?:.+\\/)?';
-const TO_ESCAPE = [
+/** @type {?} */
+const TO_ESCAPE_BASE = [
     { replace: /\./g, with: '\\.' },
-    { replace: /\?/g, with: '\\?' },
     { replace: /\+/g, with: '\\+' },
     { replace: /\*/g, with: WILD_SINGLE },
 ];
+/** @type {?} */
+const TO_ESCAPE_WILDCARD_QM = [
+    ...TO_ESCAPE_BASE,
+    { replace: /\?/g, with: QUESTION_MARK },
+];
+/** @type {?} */
+const TO_ESCAPE_LITERAL_QM = [
+    ...TO_ESCAPE_BASE,
+    { replace: /\?/g, with: '\\?' },
+];
 /**
  * @param {?} glob
+ * @param {?=} literalQuestionMark
  * @return {?}
  */
-function globToRegex(glob) {
-    const /** @type {?} */ segments = glob.split('/').reverse();
-    let /** @type {?} */ regex = '';
+function globToRegex(glob, literalQuestionMark = false) {
+    /** @type {?} */
+    const toEscape = literalQuestionMark ? TO_ESCAPE_LITERAL_QM : TO_ESCAPE_WILDCARD_QM;
+    /** @type {?} */
+    const segments = glob.split('/').reverse();
+    /** @type {?} */
+    let regex = '';
     while (segments.length > 0) {
-        const /** @type {?} */ segment = /** @type {?} */ ((segments.pop()));
+        /** @type {?} */
+        const segment = /** @type {?} */ ((segments.pop()));
         if (segment === '**') {
             if (segments.length > 0) {
                 regex += WILD_OPEN;
@@ -94,7 +122,8 @@ function globToRegex(glob) {
             }
         }
         else {
-            const /** @type {?} */ processed = TO_ESCAPE.reduce((segment, escape) => segment.replace(escape.replace, escape.with), segment);
+            /** @type {?} */
+            const processed = toEscape.reduce((segment, escape) => segment.replace(escape.replace, escape.with), segment);
             regex += processed;
             if (segments.length > 0) {
                 regex += '\\/';
@@ -104,17 +133,9 @@ function globToRegex(glob) {
     return regex;
 }
 
-var __awaiter = (undefined && undefined.__awaiter) || function (thisArg, _arguments, P, generator) {
-    return new (P || (P = Promise))(function (resolve, reject) {
-        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
-        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
-        step((generator = generator.apply(thisArg, _arguments || [])).next());
-    });
-};
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes} checked by tsc
+ * @suppress {checkTypes,extraRequire,uselessCode} checked by tsc
  */
 /**
  * @license
@@ -123,6 +144,13 @@ var __awaiter = (undefined && undefined.__awaiter) || function (thisArg, _argume
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
+/** @type {?} */
+const DEFAULT_NAVIGATION_URLS = [
+    '/**',
+    '!/**/*.*',
+    '!/**/*__*',
+    '!/**/*__*/**',
+];
 /**
  * Consumes service worker configuration files and processes them into control files.
  *
@@ -143,13 +171,17 @@ class Generator {
      */
     process(config) {
         return __awaiter(this, void 0, void 0, function* () {
-            const /** @type {?} */ hashTable = {};
+            /** @type {?} */
+            const unorderedHashTable = {};
+            /** @type {?} */
+            const assetGroups = yield this.processAssetGroups(config, unorderedHashTable);
             return {
                 configVersion: 1,
-                index: joinUrls(this.baseHref, config.index),
                 appData: config.appData,
-                assetGroups: yield this.processAssetGroups(config, hashTable),
-                dataGroups: this.processDataGroups(config), hashTable,
+                index: joinUrls(this.baseHref, config.index), assetGroups,
+                dataGroups: this.processDataGroups(config),
+                hashTable: withOrderedKeys(unorderedHashTable),
+                navigationUrls: processNavigationUrls(this.baseHref, config.navigationUrls),
             };
         });
     }
@@ -160,36 +192,40 @@ class Generator {
      */
     processAssetGroups(config, hashTable) {
         return __awaiter(this, void 0, void 0, function* () {
-            const /** @type {?} */ seenMap = new Set();
+            /** @type {?} */
+            const seenMap = new Set();
             return Promise.all((config.assetGroups || []).map((group) => __awaiter(this, void 0, void 0, function* () {
-                const /** @type {?} */ fileMatcher = globListToMatcher(group.resources.files || []);
-                const /** @type {?} */ versionedMatcher = globListToMatcher(group.resources.versionedFiles || []);
-                const /** @type {?} */ allFiles = (yield this.fs.list('/'));
-                const /** @type {?} */ versionedFiles = allFiles.filter(versionedMatcher).filter(file => !seenMap.has(file));
-                versionedFiles.forEach(file => seenMap.add(file));
-                const /** @type {?} */ plainFiles = allFiles.filter(fileMatcher).filter(file => !seenMap.has(file));
+                if (group.resources.versionedFiles) {
+                    console.warn(`Asset-group '${group.name}' in 'ngsw-config.json' uses the 'versionedFiles' option.\n` +
+                        'As of v6 \'versionedFiles\' and \'files\' options have the same behavior. ' +
+                        'Use \'files\' instead.');
+                }
+                /** @type {?} */
+                const fileMatcher = globListToMatcher(group.resources.files || []);
+                /** @type {?} */
+                const versionedMatcher = globListToMatcher(group.resources.versionedFiles || []);
+                /** @type {?} */
+                const allFiles = yield this.fs.list('/');
+                /** @type {?} */
+                const plainFiles = allFiles.filter(fileMatcher).filter(file => !seenMap.has(file));
                 plainFiles.forEach(file => seenMap.add(file));
-                // Add the hashes.
-                yield [...versionedFiles, ...plainFiles].reduce((previous, file) => __awaiter(this, void 0, void 0, function* () {
+                /** @type {?} */
+                const versionedFiles = allFiles.filter(versionedMatcher).filter(file => !seenMap.has(file));
+                versionedFiles.forEach(file => seenMap.add(file));
+                /** @type {?} */
+                const matchedFiles = [...plainFiles, ...versionedFiles].sort();
+                yield matchedFiles.reduce((previous, file) => __awaiter(this, void 0, void 0, function* () {
                     yield previous;
-                    const /** @type {?} */ hash = yield this.fs.hash(file);
+                    /** @type {?} */
+                    const hash = yield this.fs.hash(file);
                     hashTable[joinUrls(this.baseHref, file)] = hash;
                 }), Promise.resolve());
-                // Figure out the patterns.
-                const /** @type {?} */ patterns = (group.resources.urls || [])
-                    .map(glob => glob.startsWith('/') || glob.indexOf('://') !== -1 ?
-                    glob :
-                    joinUrls(this.baseHref, glob))
-                    .map(glob => globToRegex(glob));
                 return {
                     name: group.name,
                     installMode: group.installMode || 'prefetch',
                     updateMode: group.updateMode || group.installMode || 'prefetch',
-                    urls: (/** @type {?} */ ([]))
-                        .concat(plainFiles)
-                        .concat(versionedFiles)
-                        .map(url => joinUrls(this.baseHref, url)),
-                    patterns,
+                    urls: matchedFiles.map(url => joinUrls(this.baseHref, url)),
+                    patterns: (group.resources.urls || []).map(url => urlToRegex(url, this.baseHref, true)),
                 };
             })));
         });
@@ -200,14 +236,9 @@ class Generator {
      */
     processDataGroups(config) {
         return (config.dataGroups || []).map(group => {
-            const /** @type {?} */ patterns = group.urls
-                .map(glob => glob.startsWith('/') || glob.indexOf('://') !== -1 ?
-                glob :
-                joinUrls(this.baseHref, glob))
-                .map(glob => globToRegex(glob));
             return {
                 name: group.name,
-                patterns,
+                patterns: group.urls.map(url => urlToRegex(url, this.baseHref, true)),
                 strategy: group.cacheConfig.strategy || 'performance',
                 maxSize: group.cacheConfig.maxSize,
                 maxAge: parseDurationToMs(group.cacheConfig.maxAge),
@@ -218,11 +249,25 @@ class Generator {
     }
 }
 /**
+ * @param {?} baseHref
+ * @param {?=} urls
+ * @return {?}
+ */
+function processNavigationUrls(baseHref, urls = DEFAULT_NAVIGATION_URLS) {
+    return urls.map(url => {
+        /** @type {?} */
+        const positive = !url.startsWith('!');
+        url = positive ? url : url.substr(1);
+        return { positive, regex: `^${urlToRegex(url, baseHref)}$` };
+    });
+}
+/**
  * @param {?} globs
  * @return {?}
  */
 function globListToMatcher(globs) {
-    const /** @type {?} */ patterns = globs.map(pattern => {
+    /** @type {?} */
+    const patterns = globs.map(pattern => {
         if (pattern.startsWith('!')) {
             return {
                 positive: false,
@@ -244,7 +289,8 @@ function globListToMatcher(globs) {
  * @return {?}
  */
 function matches(file, patterns) {
-    const /** @type {?} */ res = patterns.reduce((isMatch, pattern) => {
+    /** @type {?} */
+    const res = patterns.reduce((isMatch, pattern) => {
         if (pattern.positive) {
             return isMatch || pattern.regex.test(file);
         }
@@ -253,6 +299,18 @@ function matches(file, patterns) {
         }
     }, false);
     return res;
+}
+/**
+ * @param {?} url
+ * @param {?} baseHref
+ * @param {?=} literalQuestionMark
+ * @return {?}
+ */
+function urlToRegex(url, baseHref, literalQuestionMark) {
+    if (!url.startsWith('/') && url.indexOf('://') === -1) {
+        url = joinUrls(baseHref, url);
+    }
+    return globToRegex(url, literalQuestionMark);
 }
 /**
  * @param {?} a
@@ -268,15 +326,33 @@ function joinUrls(a, b) {
     }
     return a + b;
 }
+/**
+ * @template T
+ * @param {?} unorderedObj
+ * @return {?}
+ */
+function withOrderedKeys(unorderedObj) {
+    /** @type {?} */
+    const orderedObj = /** @type {?} */ ({});
+    Object.keys(unorderedObj).sort().forEach(key => orderedObj[key] = unorderedObj[key]);
+    return orderedObj;
+}
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes} checked by tsc
+ * @suppress {checkTypes,extraRequire,uselessCode} checked by tsc
+ */
+/**
+ * @license
+ * Copyright Google Inc. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
  */
 
 /**
  * @fileoverview added by tsickle
- * @suppress {checkTypes} checked by tsc
+ * @suppress {checkTypes,extraRequire,uselessCode} checked by tsc
  */
 /**
  * Generated bundle index. Do not edit.
