@@ -3,4 +3,4 @@
  */
 export * from './public_api';
 export { NgswCommChannel as ɵe } from './src/low_level';
-export { OPTS as ɵb, SCRIPT as ɵa, ngswAppInitializer as ɵc, ngswCommChannelFactory as ɵd } from './src/module';
+export { RegistrationOptions as ɵa, SCRIPT as ɵb, ngswAppInitializer as ɵc, ngswCommChannelFactory as ɵd } from './src/module';
