@@ -7,13 +7,25 @@
  */
 import { InjectionToken, Injector, ModuleWithProviders } from '@angular/core';
 import { NgswCommChannel } from './low_level';
+export declare abstract class RegistrationOptions {
+    scope?: string;
+    enabled?: boolean;
+}
 export declare const SCRIPT: InjectionToken<string>;
-export declare const OPTS: InjectionToken<Object>;
-export declare function ngswAppInitializer(injector: Injector, script: string, options: RegistrationOptions): Function;
-export declare function ngswCommChannelFactory(): NgswCommChannel;
+export declare function ngswAppInitializer(injector: Injector, script: string, options: RegistrationOptions, platformId: string): Function;
+export declare function ngswCommChannelFactory(opts: RegistrationOptions, platformId: string): NgswCommChannel;
 /**
  * @experimental
  */
 export declare class ServiceWorkerModule {
-    static register(script: string, opts?: RegistrationOptions): ModuleWithProviders;
+    /**
+     * Register the given Angular Service Worker script.
+     *
+     * If `enabled` is set to `false` in the given options, the module will behave as if service
+     * workers are not supported by the browser, and the service worker will not be registered.
+     */
+    static register(script: string, opts?: {
+        scope?: string;
+        enabled?: boolean;
+    }): ModuleWithProviders;
 }
